==> source_code/notebooks/5_global_groups_unite.ipynb <==
import os
import pandas as pd

dataset = {
    0 : "absorption",
    1 : "enantioselectivity",
    2 : "localization",
    3 : "T50"
}
method = {
    0:"FFT",
    1:"NLP",
    2:"Properties"
}
bioembedding = {
    0:"bepler",
    1:"esm",
    2:"fasttext",
    3:"plus_rnn",
    4:"prottrans"
}
distances = {
    1 : "Euclidean",
    2 : "Braycurtis",
    3 : "Canberra",
    4 : "Chebyshev",
    5 : "Cityblock",
    6 : "Correlation",
    7 : "Cosine",
    8 : "Minkowski",
    9 : "Hamming"
}
absorption_parent = {
    0 : "sequence_40826",
}
localization_parent = {
    0 : "sequence_51016",
    1 : "sequence_51170",
    2 : "sequence_32130",
}
enantio_parent = {
    0 : "sequence_47787",
}
t50_parent = {
    0 : "sequence_53082",
    1 : "sequence_19402",
    2 : "sequence_20054",
}

for a in range (0,4): #Se elije el dataset
    for b in range(0,3): #Se elije el metodo de representacion numerica
        if b==1: #Si es NLP entra a este if
            for c in range(0,5): #Cual de los 5 metodos de PLM
                for d in range(1,10): #Se elije la distancia
                    df_profiles = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/"+bioembedding[c]+"/"+distances[d]+"/generated_profiles.csv")
                    df_distances = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/"+bioembedding[c]+"/"+distances[d]+"/distance_estimator.csv")
                    profiles = df_profiles.profile.unique()
                    Cluster = pd.DataFrame()
                    for f in range(0,len(profiles)): #para entrar a cada perfil
                        df_f = (df_profiles["profile"] == profiles[f])
                        clust_aux = df_profiles[df_f]
                        if a == 0: #Absorption
                            for g in range(0,len(clust_aux.index)): #Esto para elegir la o las nativas
                                df_filter = df_distances[(df_distances.id_seq1==absorption_parent[0]) & (df_distances.id_seq2 == clust_aux.iloc[g][0])]
                                df_filter.insert(5,"profile",profiles[f],allow_duplicates=False)
                                Cluster = pd.concat([Cluster,df_filter])
                        if a == 1: #Enantioselectivity
                            for g in range(0,len(clust_aux.index)): #Esto para elegir la o las nativas
                                df_filter = df_distances[(df_distances.id_seq1==enantio_parent[0]) & (df_distances.id_seq2 == clust_aux.iloc[g][0])]
                                df_filter.insert(5,"profile",profiles[f],allow_duplicates=False)
                                Cluster = pd.concat([Cluster,df_filter])
                        if a == 2: #Localization
                            for parent in range(0,3):
                                for g in range(0,len(clust_aux.index)): #Esto para elegir la o las nativas
                                    df_filter = df_distances[(df_distances.id_seq1==localization_parent[parent]) & (df_distances.id_seq2 == clust_aux.iloc[g][0])]
                                    df_filter.insert(5,"profile",profiles[f],allow_duplicates=False)
                                    Cluster = pd.concat([Cluster,df_filter])
                        if a == 3: #T50
                            for parent in range(0,3):
                                for g in range(0,len(clust_aux.index)): #Esto para elegir la o las nativas
                                    df_filter = df_distances[(df_distances.id_seq1==t50_parent[parent]) & (df_distances.id_seq2 == clust_aux.iloc[g][0])]
                                    df_filter.insert(5,"profile",profiles[f],allow_duplicates=False)
                                    Cluster = pd.concat([Cluster,df_filter])
                    Cluster.to_csv("../../results_demo/"+dataset[a]+"/NLP/"+bioembedding[c]+"/"+distances[d]+"/cluster_unite_without.csv")
                    print("Cluster unite for "+dataset[a]+"/NLP/"+bioembedding[c]+"/"+distances[d])
        if b == 0:
            for c in range(0,8):
                for d in range(1,10): #Se elije la distancia
                    df_profiles = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+"/generated_profiles.csv")
                    df_distances = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+"/distance_estimator.csv")
                    profiles = df_profiles.profile.unique()
                    Cluster = pd.DataFrame()
                    for f in range(0,len(profiles)): #para entrar a cada perfil
                        df_f = (df_profiles["profile"] == profiles[f])
                        clust_aux = df_profiles[df_f]
                        if a == 0: #Absorption
                            for g in range(0,len(clust_aux.index)): #Esto para elegir la o las nativas
                                df_filter = df_distances[(df_distances.id_seq1==absorption_parent[0]) & (df_distances.id_seq2 == clust_aux.iloc[g][0])]
                                df_filter.insert(5,"profile",profiles[f],allow_duplicates=False)
                                Cluster = pd.concat([Cluster,df_filter])
                        if a == 1: #Enantioselectivity
                            for g in range(0,len(clust_aux.index)): #Esto para elegir la o las nativas
                                df_filter = df_distances[(df_distances.id_seq1==enantio_parent[0]) & (df_distances.id_seq2 == clust_aux.iloc[g][0])]
                                df_filter.insert(5,"profile",profiles[f],allow_duplicates=False)
                                Cluster = pd.concat([Cluster,df_filter])
                        if a == 2: #Localization
                            for parent in range(0,3):
                                for g in range(0,len(clust_aux.index)): #Esto para elegir la o las nativas
                                    df_filter = df_distances[(df_distances.id_seq1==localization_parent[parent]) & (df_distances.id_seq2 == clust_aux.iloc[g][0])]
                                    df_filter.insert(5,"profile",profiles[f],allow_duplicates=False)
                                    Cluster = pd.concat([Cluster,df_filter])
                        if a == 3: #T50
                            for parent in range(0,3):
                                for g in range(0,len(clust_aux.index)): #Esto para elegir la o las nativas
                                    df_filter = df_distances[(df_distances.id_seq1==t50_parent[parent]) & (df_distances.id_seq2 == clust_aux.iloc[g][0])]
                                    df_filter.insert(5,"profile",profiles[f],allow_duplicates=False)
                                    Cluster = pd.concat([Cluster,df_filter])
                    Cluster.to_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+"/cluster_unite_without.csv")
                    print("Cluster unite for "+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d])
        if b == 2:
            for c in range(0,8):
                for d in range(1,10): #Se elije la distancia
                    df_profiles = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+"/generated_profiles.csv")
                    df_distances = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+"/distance_estimator.csv")
                    profiles = df_profiles.profile.unique()
                    Cluster = pd.DataFrame()
                    for f in range(0,len(profiles)): #para entrar a cada perfil
                        df_f = (df_profiles["profile"] == profiles[f])
                        clust_aux = df_profiles[df_f]
                        if a == 0: #Absorption
                            for g in range(0,len(clust_aux.index)): #Esto para elegir la o las nativas
                                df_filter = df_distances[(df_distances.id_seq1==absorption_parent[0]) & (df_distances.id_seq2 == clust_aux.iloc[g][0])]
                                df_filter.insert(5,"profile",profiles[f],allow_duplicates=False)
                                Cluster = pd.concat([Cluster,df_filter])
                        if a == 1: #Enantioselectivity
                            for g in range(0,len(clust_aux.index)): #Esto para elegir la o las nativas
                                df_filter = df_distances[(df_distances.id_seq1==enantio_parent[0]) & (df_distances.id_seq2 == clust_aux.iloc[g][0])]
                                df_filter.insert(5,"profile",profiles[f],allow_duplicates=False)
                                Cluster = pd.concat([Cluster,df_filter])
                        if a == 2: #Localization
                            for parent in range(0,3):
                                for g in range(0,len(clust_aux.index)): #Esto para elegir la o las nativas
                                    df_filter = df_distances[(df_distances.id_seq1==localization_parent[parent]) & (df_distances.id_seq2 == clust_aux.iloc[g][0])]
                                    df_filter.insert(5,"profile",profiles[f],allow_duplicates=False)
                                    Cluster = pd.concat([Cluster,df_filter])
                        if a == 3: #T50
                            for parent in range(0,3):
                                for g in range(0,len(clust_aux.index)): #Esto para elegir la o las nativas
                                    df_filter = df_distances[(df_distances.id_seq1==t50_parent[parent]) & (df_distances.id_seq2 == clust_aux.iloc[g][0])]
                                    df_filter.insert(5,"profile",profiles[f],allow_duplicates=False)
                                    Cluster = pd.concat([Cluster,df_filter])
                    Cluster.to_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+"/cluster_unite_without.csv")
                    print("Cluster unite for "+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d])
